+++ app/Dictionary.txt
@@ -1,2 +1,2 @@
-VAB:Virginia Beach
-Area_Occ:Location
+Virginia Beach:VAB
+Location:Area_Occ
